refactor: move deepCopy function to converterUtils.js
